requirements.txt
asgiref==3.2.7
Django==3.0.5
django-extensions==2.2.9
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.2.7
get==2019.4.13
pylint-django==2.0.15
post==2019.4.13
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
requests==2.21.0
sqlparse==0.3.1